* 실행 - 저장 - 종료까지 존재하던 오류 제거

+++ trunk/tester/Form.Setting.pas
@@ -7,9 +7,12 @@ uses
   System.Classes, Vcl.Graphics, System.UITypes,
   Vcl.Controls, Vcl.Forms, Vcl.Dialogs, Vcl.StdCtrls, Generics.Collections,
   uFileFunctions, uSaveFile, uGSTestThread, Device.PhysicalDrive,
-  uPartitionListGetter, Vcl.ComCtrls, Threading;
+  uPartitionListGetter, Vcl.ComCtrls, uAutoPhysicalDriveListGetter,
+  uPhysicalDriveList;
 
 type
+  EDriveNotFound = class(EResNotFound);
+
   TfSetting = class(TForm)
     GroupBox1: TGroupBox;
     GroupBox2: TGroupBox;
@@ -40,6 +43,19 @@ type
     FSavePath: String;
     FTracePath: String;
     FNeedToLoad: Boolean;
+    function GetSavePath: Boolean;
+    procedure InsertDriveByNumber(SaveFile: TSaveFile);
+    procedure InsertDriveByModelSerial(SaveFile: TSaveFile);
+    procedure SetFormBySaveFile(SaveFile: TSaveFile);
+    function IsAllOptionSet: Boolean;
+    function IsDestinationSet: Boolean;
+    function IsFFRSet: Boolean;
+    function IsMaxTBWBiggerThanRetentionTBW: Boolean;
+    function IsTBWToRetentionSet: Boolean;
+    function IsTBWToWriteSet: Boolean;
+    procedure SetTracePath;
+    procedure IfAlertExistsDelete;
+    procedure SetSavePath;
   public
     property SavePath: String read FSavePath;
     property TracePath: String read FTracePath;
@@ -60,11 +76,8 @@ procedure TfSetting.bOpenExistClick(Sender: TObject);
 var
   SaveFile: TSaveFile;
 begin
-  FSavePath := SelectDirectory('로그가 저장된 폴더를 선택해주세요', AppPath);
-
-  if FSavePath = '' then
+  if not GetSavePath then
     exit;
-
   if FileExists(FSavePath + 'settings.ini') = false then
   begin
     ShowMessage('테스트 파일이 없습니다.');
@@ -73,60 +86,65 @@ begin
 
   SaveFile := TSaveFile.Create;
   SaveFile.LoadFromFile(FSavePath + 'settings.ini');
-
-  if FDriveList.IndexOf(SaveFile.Disknum) >= 0 then
-  begin
-    FDriveList.Insert(0, SaveFile.Disknum);
-    cDestination.Items.Insert(0, 'Open');
-    cDestination.ItemIndex := 0;
-  end
-  else
-  begin
-    FDriveList.Insert(0, FindDrive(SaveFile.Model,
-                                   SaveFile.Serial));
-    cDestination.Items.Insert(0, 'Open');
-    cDestination.ItemIndex := 0;
-  end;
-
-  eDestTBW.Text := IntToStr(SaveFile.MaxTBW shr ByteToTB);
-  eRetentionTBW.Text := IntToStr(SaveFile.RetTBW shr ByteToTB);
+  SetFormBySaveFile(SaveFile);
   FNeedToLoad := true;
-
   FOptionsSet := true;
   Close;
 end;
 
-procedure TfSetting.bStartNewClick(Sender: TObject);
-var
-  MaxTBW, MaxReten: Integer;
-  MaxFFR: Integer;
+function TfSetting.IsDestinationSet: Boolean;
 begin
-  if cDestination.ItemIndex = -1 then
-  begin
+  result := cDestination.ItemIndex <> -1;
+  if not result then
     ShowMessage('대상 위치를 올바르게 입력해주세요');
-    exit;
-  end;
-  if TryStrToInt(eDestTBW.Text, MaxTBW) = false then
-  begin
-    ShowMessage('목표 TBW를 올바르게 입력해주세요');
-    exit;
-  end;
-  if TryStrToInt(eRetentionTBW.Text, MaxReten) = false then
-  begin
-    ShowMessage('리텐션 테스트 주기를 올바르게 입력해주세요');
-    exit;
-  end;
-  if TryStrToInt(eFFR.Text, MaxFFR) = false then
-  begin
-    ShowMessage('기능 실패율을 올바르게 입력해주세요');
-    exit;
-  end;
-  if MaxTBW < MaxReten then
-  begin
-    ShowMessage('목표 TBW는 리텐션 테스트 주기보다 작을 수 없습니다');
-    exit;
-  end;
+end;
 
+function TfSetting.IsTBWToWriteSet: Boolean;
+var
+  Dummy: Integer;
+begin
+  result := TryStrToInt(eDestTBW.Text, Dummy);
+  if not result then
+    ShowMessage('목표 TBW를 올바르게 입력해주세요');
+end;
+
+function TfSetting.IsTBWToRetentionSet: Boolean;
+var
+  Dummy: Integer;
+begin
+  result := TryStrToInt(eRetentionTBW.Text, Dummy);
+  if not result then
+    ShowMessage('리텐션 테스트 주기를 올바르게 입력해주세요');
+end;
+
+function TfSetting.IsFFRSet: Boolean;
+var
+  Dummy: Integer;
+begin
+  result := TryStrToInt(eFFR.Text, Dummy);
+  if not result then
+    ShowMessage('기능 실패율을 올바르게 입력해주세요');
+end;
+
+function TfSetting.IsMaxTBWBiggerThanRetentionTBW: Boolean;
+begin
+  result := StrToInt(eDestTBW.Text) < StrToInt(eRetentionTBW.Text);
+  if not result then
+    ShowMessage('목표 TBW는 리텐션 테스트 주기보다 작을 수 없습니다');
+end;
+
+function TfSetting.IsAllOptionSet: Boolean;
+begin
+  result :=
+    IsDestinationSet and
+    IsTBWToWriteSet and
+    IsTBWToRetentionSet and
+    IsFFRSet and
+    IsMaxTBWBiggerThanRetentionTBW;
+end;
+
+procedure TfSetting.SetTracePath;
+begin
   FTracePath := AppPath + 'mt.txt';
   while (FTracePath = '') or (not(FileExists(FTracePath))) do
   begin
@@ -135,7 +153,16 @@ begin
 
     FTracePath := oTrace.FileName;
   end;
+end;
 
+procedure TfSetting.SetSavePath;
+  function IsUserNeedOverwrite: Boolean;
+  begin
+    result := MessageDlg(
+      '해당 폴더에 이미 로그가 있습니다. 덮어씌우시겠습니까?', mtWarning,
+      mbOKCancel, 0) = mrOk;
+  end;
+begin
   repeat
     FSavePath := SelectDirectory('로그가 저장될 폴더를 선택해주세요', AppPath);
     if FSavePath = '' then
@@ -143,18 +170,28 @@ begin
 
     if FileExists(FSavePath + 'settings.ini') then
     begin
-      if MessageDlg('해당 폴더에 이미 로그가 있습니다. 덮어씌우시겠습니까?',
-                    mtWarning, mbOKCancel, 0) = mrCancel then
-        Exit
+      if IsUserNeedOverwrite then
+        DeleteFile(FSavePath + 'settings.ini')
       else
       begin
-        DeleteFile(FSavePath + 'settings.ini');
+        FSavePath := '';
+        exit;
       end;
     end;
-  until (FSavePath <> '') and (not(FileExists(FSavePath + 'settings.ini')));
+  until FSavePath <> '';
+end;
 
-  if FileExists(FSavePath + 'alert.txt') then
-    DeleteFile(PChar(FSavePath + 'alert.txt'));
+procedure TfSetting.bStartNewClick(Sender: TObject);
+begin
+  SetTracePath;
+  if FTracePath = '' then
+    exit;
+
+  SetSavePath;
+  if FSavePath = '' then
+    exit;
+
+  IfAlertExistsDelete;
 
   FOptionsSet := true;
   Close;
@@ -194,78 +231,88 @@ begin
   exit(FDriveList[cDestination.ItemIndex]);
 end;
 
+procedure TfSetting.IfAlertExistsDelete;
+begin
+  if FileExists(FSavePath + 'alert.txt') then
+    DeleteFile(PChar(FSavePath + 'alert.txt'));
+end;
+
+procedure TfSetting.SetFormBySaveFile(SaveFile: TSaveFile);
+begin
+  if FDriveList.IndexOf(SaveFile.Disknum) >= 0 then
+    InsertDriveByNumber(SaveFile)
+  else
+    InsertDriveByModelSerial(SaveFile);
+  eDestTBW.Text := IntToStr(SaveFile.MaxTBW shr ByteToTB);
+  eRetentionTBW.Text := IntToStr(SaveFile.RetTBW shr ByteToTB);
+end;
+
+procedure TfSetting.InsertDriveByNumber(SaveFile: TSaveFile);
+begin
+  FDriveList.Insert(0, SaveFile.Disknum);
+  cDestination.Items.Insert(0, 'Open');
+  cDestination.ItemIndex := 0;
+end;
+
+procedure TfSetting.InsertDriveByModelSerial(SaveFile: TSaveFile);
+begin
+  FDriveList.Insert(0, FindDrive(SaveFile.Model, SaveFile.Serial));
+  cDestination.Items.Insert(0, 'Open');
+  cDestination.ItemIndex := 0;
+end;
+
+function TfSetting.GetSavePath: Boolean;
+begin
+  FSavePath := SelectDirectory('로그가 저장된 폴더를 선택해주세요', AppPath);
+  result := FSavePath <> '';
+end;
+
 function TfSetting.FindDrive(Model, Serial: String): Integer;
 var
+  DriveList: TPhysicalDriveList;
   PhysicalDrive: IPhysicalDrive;
-  CurrDrv: Integer;
-  PartitionList: TPartitionList;
 begin
+  DriveList := AutoPhysicalDriveListGetter.GetPhysicalDriveList;
   result := -1;
-  for CurrDrv := 0 to 99 do
+  for PhysicalDrive in DriveList do
   begin
-    try
-      PhysicalDrive := TPhysicalDrive.Create(
-        TPhysicalDrive.BuildFileAddressByNumber(CurrDrv));
-
-      if not PhysicalDrive.IsDriveAvailable then
-        Continue;
-    except
-      Continue;
-    end;
-
-    PartitionList := PhysicalDrive.GetPartitionList;
-    if PartitionList.Count > 0 then
-    begin
-      FreeAndNil(PartitionList);
-      Continue;
-    end;
-
     if (Model = PhysicalDrive.IdentifyDeviceResult.Model) and
        (Serial = PhysicalDrive.IdentifyDeviceResult.Serial) then
-    begin
-      result := CurrDrv;
-      break;
-    end;
-    FreeAndNil(PartitionList);
+      result :=
+        StrToInt(PhysicalDrive.GetPathOfFileAccessingWithoutPrefix);
   end;
+  if result = -1 then
+    raise
+      EDriveNotFound.Create('DriveNotFound Model ' + Model + ', Serial ' +
+        Serial);
+  FreeAndNil(DriveList);
 end;
 
 procedure TfSetting.RefreshDrives;
+var
+  DriveList: TPhysicalDriveList;
+  PartitionList: TPartitionList;
+  PhysicalDrive: IPhysicalDrive;
+  CurrentDriveNumber: Integer;
 begin
-  TParallel.For(0, 99, procedure (CurrDrv: Integer)
-  var
-    PhysicalDrive: IPhysicalDrive;
-    PartitionList: TPartitionList;
+  DriveList := AutoPhysicalDriveListGetter.GetPhysicalDriveList;
+  for PhysicalDrive in DriveList do
   begin
-    try
-      PhysicalDrive := TPhysicalDrive.Create(
-        TPhysicalDrive.BuildFileAddressByNumber(CurrDrv));
-
-      if not PhysicalDrive.IsDriveAvailable then
-        exit;
-    except
-      exit;
-    end;
-
     PartitionList := PhysicalDrive.GetPartitionList;
-    if PartitionList.Count > 0 then
+    if PartitionList.Count = 0 then
     begin
-      FreeAndNil(PartitionList);
-      exit;
-    end;
-
-    TThread.Queue(TThread.CurrentThread, procedure
-    begin
-      FDriveList.Add(CurrDrv);
-      cDestination.Items.Add(IntToStr(CurrDrv) + ' - ' +
+      CurrentDriveNumber :=
+        StrToInt(PhysicalDrive.GetPathOfFileAccessingWithoutPrefix);
+      FDriveList.Add(CurrentDriveNumber);
+      cDestination.Items.Add(
+        PhysicalDrive.GetPathOfFileAccessingWithoutPrefix + ' - ' +
         PhysicalDrive.IdentifyDeviceResult.Model);
       if cDestination.ItemIndex = -1 then
         cDestination.ItemIndex := 0;
-    end);
-
+    end;
     FreeAndNil(PartitionList);
-  end);
+  end;
+  FreeAndNil(DriveList);
 end;
-
 
 end.
